requirements.txt
alembic==1.5.4
atomicwrites==1.4.0
attrs==20.3.0
cffi==1.14.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typing-extensions==3.7.4.3
webargs==7.0.1
Werkzeug==1.0.1
zipp==3.4.0
